New branch with UUID hkfzshafy7b5p1zthbjvanydm
A utility for reversing the order of bytes in a file shall be developed,
i. e. something like "rev", but byte- rather than line-oriented.

The utility shall work with seekable input streams of unrestricted size,
and with non-seekable input streams as long as they are not longer than
the size of an internal fixed-sized buffer (typically a couple of
megabytes; i. e. not too large but large enough for most text files.)